fix: add long rpc message to example
