+++ src/room/rpc/utils.ts
